Fixed many many many issues
