+++ application_details.json
@@ -3,6 +3,6 @@
         "nest": "Discord Bot"
     },
     "753668246972137513": {
-        "nest": "Discord Bot"
+        "subtype": "Python"
     }
 }
